CY-698 Make removing temp files more robust

+++ cosmo_tester/framework/test_hosts.py
@@ -163,23 +163,27 @@ class VM(object):
     def get_remote_file_content(self, remote_path, use_sudo=True):
         tmp_local_path = os.path.join(self._tmpdir, str(uuid.uuid4()))
 
-        self.get_remote_file(remote_path, tmp_local_path, use_sudo)
-
-        with open(tmp_local_path, 'r') as f:
-            content = f.read()
-
-        os.unlink(tmp_local_path)
+        try:
+            self.get_remote_file(remote_path, tmp_local_path, use_sudo)
+            with open(tmp_local_path, 'r') as f:
+                content = f.read()
+        finally:
+            if os.path.exists(tmp_local_path):
+                os.unlink(tmp_local_path)
         return content
 
     def put_remote_file_content(self, remote_path, content, use_sudo=True):
         tmp_local_path = os.path.join(self._tmpdir, str(uuid.uuid4()))
 
-        with open(tmp_local_path, 'w') as f:
-            f.write(content)
+        try:
+            with open(tmp_local_path, 'w') as f:
+                f.write(content)
 
-        self.put_remote_file(remote_path, tmp_local_path, use_sudo)
+            self.put_remote_file(remote_path, tmp_local_path, use_sudo)
 
-        os.unlink(tmp_local_path)
+        finally:
+            if os.path.exists(tmp_local_path):
+                os.unlink(tmp_local_path)
 
     def run_command(self, command, use_sudo=False):
         with self.ssh() as fabric_ssh:
